taxon_search v0.0.1 TaxonSearch

Provides a function TaxonSearch.get_species_names/1 to get species names associated with the specified common_name.

Functions

get_species_names(common_name)

Gets species names associated with the common_name.

Parameters

common_name: String that represents the common name of the species you’re interested in

Examples

iex> TaxonSearch.get_species_names("scaly-headed pionus")
["Pionus maximiliani"]

iex> TaxonSearch.get_species_names("Senegal parrot")
["Poicephalus senegalus", "Poicephalus versteri"]